+++ src/main/java/com/landawn/abacus/da/hbase/AnyGet.java
@@ -35,7 +35,7 @@ import org.apache.hadoop.hbase.io.TimeRange;
  * @see org.apache.hadoop.hbase.client.Get
  * @since 0.8
  */
-public final class AnyGet extends AnyQuery<AnyGet> implements Comparable<Row> {
+public final class AnyGet extends AnyQuery<AnyGet> implements Row {
 
     /** The get. */
     private final Get get;
@@ -45,7 +45,7 @@ public final class AnyGet extends AnyQuery<AnyGet> implements Comparable<Row> {
      *
      * @param rowKey
      */
-      AnyGet(Object rowKey) {
+    AnyGet(Object rowKey) {
         super(new Get(toRowKeyBytes(rowKey)));
         this.get = (Get) query;
     }
@@ -57,7 +57,7 @@ public final class AnyGet extends AnyQuery<AnyGet> implements Comparable<Row> {
      * @param rowOffset
      * @param rowLength
      */
-      AnyGet(Object rowKey, int rowOffset, int rowLength) {
+    AnyGet(Object rowKey, int rowOffset, int rowLength) {
         super(new Get(toRowKeyBytes(rowKey), rowOffset, rowLength));
         this.get = (Get) query;
     }
@@ -67,7 +67,7 @@ public final class AnyGet extends AnyQuery<AnyGet> implements Comparable<Row> {
      *
      * @param rowKey
      */
-      AnyGet(ByteBuffer rowKey) {
+    AnyGet(ByteBuffer rowKey) {
         super(new Get(rowKey));
         this.get = (Get) query;
     }
@@ -77,7 +77,7 @@ public final class AnyGet extends AnyQuery<AnyGet> implements Comparable<Row> {
      *
      * @param get
      */
-      AnyGet(Get get) {
+    AnyGet(Get get) {
         super(get);
         this.get = (Get) query;
     }
@@ -416,6 +416,7 @@ public final class AnyGet extends AnyQuery<AnyGet> implements Comparable<Row> {
      *
      * @return
      */
+    @Override
     public byte[] getRow() {
         return get.getRow();
     }